requirements.txt
aiosmtplib==1.1.7
anyio==3.7.1
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.35.3
botocore==1.35.3
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-user-login-history==1.0.6
Django==5.1
django-appconf==1.0.6
django-apscheduler==0.7.0
django-auto-logout==0.5.1
django-crontab==0.7.1
django-d1db==0.0.7
django-dbbackup==4.2.1
django-environ==0.11.2
django-import-export==4.1.1
django-login-required-middleware==0.9.0
django-phonenumber-field==8.0.0
django-recaptcha==4.0.0
django-session-timeout==0.1.0
django-turnstile==0.1.0
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
fontawesomefree==6.6.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.7
importlib_metadata==8.5.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
messages==0.8.0
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
openpyxl==3.1.5
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
phonenumbers==8.13.44
pickleshare==0.7.5
pipdeptree==2.23.4
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
psycopg2==2.9.9
pytz==2024.1
pyzmq==26.2.0
redis==5.1.0
referencing==0.35.1
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.20.0
s3transfer==0.10.2
scanreq==0.0.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tablib==3.5.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
validus==0.3.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.20.2
